+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Released
 #### Added
 - Icons for compiled file.
+- Unified commission class.
 #### Fixed
 - The version is now displayed based on app config. 
 #### Changed